Implement authorization policy (#810)

+++ app/controllers/tickets_controller.rb
@@ -1,23 +1,21 @@
 class TicketsController < ApplicationController
-  before_action :set_ticket, only: %i[ show edit update destroy ]
+  wrap_parameters exclude: :tenant_id
+
+  before_action :find_ticket, only: %i[ show update destroy ]
+  before_action :find_tenant, only: %i[ create ]
+  after_action :verify_authorized, only: %i[ destroy bulk_delete ]
+  after_action :verify_policy_scoped, except: %i[ destroy bulk_delete ]
 
   def index
-    @tickets = Ticket.includes(:author, :notes, tenant: :staff)
+    @tickets = policy_scope(Ticket).includes(:author, :notes, tenant: :staff)
     @tickets.where(tenant_id: params[:tenant_id]) if params[:tenant_id]
   end
 
   def show
   end
 
-  def new
-    @ticket = Ticket.new
-  end
-
-  def edit
-  end
-
   def create
-    @ticket = Ticket.new(ticket_params.merge(author: current_user))
+    @ticket = Ticket.new(ticket_params.merge(author: current_user, tenant_id: @tenant.id))
 
     if @ticket.save
       render :show, status: :created
@@ -35,21 +33,28 @@ class TicketsController < ApplicationController
   end
 
   def destroy
+    authorize Ticket
     @ticket.destroy
     head :no_content
   end
 
   def bulk_delete
+    authorize Ticket
     Ticket.where(id: params[:ids]).destroy_all
+    head :no_content
   end
 
   private
 
-  def set_ticket
-    @ticket = Ticket.find(params[:id])
+  def find_ticket
+    @ticket = policy_scope(Ticket).find(params[:id])
+  end
+
+  def find_tenant
+    @tenant = policy_scope(Tenant).find(params[:tenant_id])
   end
 
   def ticket_params
-    params.require(:ticket).permit(:issue, :tenant_id, :status, :urgency)
+    params.require(:ticket).permit(:issue, :status, :urgency)
   end
 end
